requirements.txt
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.4
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
av==17.0.0
bitsandbytes==0.49.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
ctranslate2==4.7.1
cut-cross-entropy==25.1.1
datasets==4.3.0
davey==0.1.5
dill==0.4.0
docstring_parser==0.17.0
edge-tts==7.2.8
faster-whisper==1.2.1
filelock==3.25.2
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2025.9.0
gguf==0.18.0
h11==0.16.0
hf-xet==1.4.2
hf_transfer==0.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.8.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mistral_common==1.10.0
mpmath==1.3.0
msgspec==0.20.0
multidict==6.7.1
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.3.5
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnxruntime==1.24.4
packaging==26.0
pandas==3.0.1
peft==0.18.1
pillow==12.1.1
propcache==0.4.1
protobuf==7.34.1
psutil==7.2.2
py-cord==2.7.1
pyarrow==23.0.1
pycountry==26.2.16
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.1
pydantic_core==2.41.5
Pygments==2.20.0
PyNaCl==1.6.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
referencing==0.37.0
regex==2026.3.32
requests==2.33.0
rich==14.3.3
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.3.0
sentencepiece==0.2.1
setuptools==70.2.0
shellingham==1.5.4
six==1.17.0
sympy==1.13.1
tabulate==0.10.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.6.0+cu124
torchao==0.16.0
torchaudio==2.6.0+cu124
torchvision==0.21.0+cu124
tqdm==4.67.3
transformers==5.3.0
triton==3.2.0
trl==0.24.0
typeguard==4.5.1
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tyro==1.0.11
unsloth @ git+https://github.com/unslothai/unsloth.git@a6c1f893fc87c0973f9c32e59ca3d7d54ffb9724
unsloth_zoo==2026.3.6
urllib3==2.6.3
wheel==0.46.3
xxhash==3.6.0
yarl==1.23.0
